requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
einops==0.7.0
filelock==3.13.1
fsspec==2024.3.0
gguf==0.6.0
gitdb==4.0.11
gitpython==3.1.42
huggingface-hub==0.21.4
idna==3.6
jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
llama-cpp-python==0.2.56
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.0
pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.1.2
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
